examples/large_models/inferentia2/llama/requirements.txt
torch-neuronx
transformers-neuronx @ git+https://github.com/aws-neuron/transformers-neuronx.git
transformers
tokenizers
sentencepiece
